refactor: move roles (???) to bot
