+++ packages/bot/src/entities/roles.ts
