+++ packages/bot/src/index.ts
@@ -22,7 +22,7 @@ import {
   ModLogModule,
   MiscStuffStage,
   RobloxGamePresenceModule,
-  // RolesModule,
+  RolesModule,
   RulesModule,
   TagsModule,
   UnfurlModule,
@@ -50,7 +50,7 @@ const stages: Array<typeof Stage | Stage> = [
   InfractionsModule,
   ModLogModule,
   RobloxGamePresenceModule,
-  // RolesModule,
+  RolesModule,
   RulesModule,
   TagsModule,
   UnfurlModule,
