+++ packages/bot/src/modules/help-channels/threads/close-command.ts
@@ -11,14 +11,15 @@ import {
 import { guild } from '../../../lib/config'
 import { api } from '../../../lib/api'
 import { IGetHelpChanByChannelIdResponse } from '../../../lib/types'
+import { HelpChannel } from '../../../entities/help-channel'
 
 export async function threadCloseCommand(msg: CommandInteraction) {
-  const { data: owner } = await api.get<IGetHelpChanByChannelIdResponse>(
-    `/helpchan/${msg.channel!.id}`,
-  )
+  const channel = await HelpChannel.findOne({
+    where: { channel_id: msg.channel!.id },
+  })
 
   if (
-    (owner && owner.user_id === msg.member?.user.id) ||
+    (channel && channel.user_id === msg.member?.user.id) ||
     (typeof msg.member?.permissions !== 'string' &&
       msg.member?.permissions.has(Permissions.FLAGS.MANAGE_MESSAGES)) ||
     (!Array.isArray(msg.member?.roles) &&
@@ -32,7 +33,7 @@ export async function threadCloseCommand(msg: CommandInteraction) {
     const roleManger = msg.member?.roles as GuildMemberRoleManager
     await roleManger.remove(guild.roles.helpCooldown)
 
-    await api.delete(`/helpchan/${msg.channel!.id}`)
+    await channel!.remove()
 
     return await (msg.channel as ThreadChannel).setArchived(
       true,
