+++ packages/server/src/api/index.ts
@@ -13,24 +13,9 @@ const router = express.Router()
 import { hello } from './meta'
 
 // Help Channels
-import {
-  listHelpChannels,
-  createHelpChannel,
-  getHelpChannelByUserId,
-  getHelpChannelByChannelId,
-  deleteHelpChannel,
-} from './help-chan'
 
 // Infractions
 import { createInfraction } from './infractions'
-
-// Message Roles
-import {
-  getMessageRole,
-  createMessageRole,
-  getMessageRolesActions,
-  createMessageRolesAction,
-} from './message-roles'
 //#endregion
 
 //#region Usage
@@ -38,21 +23,8 @@ import {
 // Meta
 router.get('/', hello)
 
-// Help Channels
-router.get('/helpchan', requiresBotAuth, listHelpChannels)
-router.get('/helpchan/user/:user_id', requiresBotAuth, getHelpChannelByUserId)
-router.get('/helpchan/:channel_id', requiresBotAuth, getHelpChannelByChannelId)
-router.post('/helpchan', requiresBotAuth, createHelpChannel)
-router.delete('/helpchan/:channel_id', requiresBotAuth, deleteHelpChannel)
-
 // Infractions
 router.post('/infractions', requiresBotAuth, createInfraction)
-
-// Message Roles
-router.get('/message-roles/:id', requiresBotAuth, getMessageRole)
-router.post('/message-roles', requiresBotAuth, createMessageRole)
-router.get('/message-roles/actions', requiresBotAuth, getMessageRolesActions)
-router.post('/message-roles/actions', requiresBotAuth, createMessageRolesAction)
 
 //#endregion
