+++ packages/server/src/lib/connect-mysql.ts
@@ -4,7 +4,6 @@
 
 import { createConnection, Connection } from 'typeorm'
 import { Infractions } from '../entities/infractions'
-import { MessageRoles, MessageRolesActions } from '../entities/roles'
 import { AccessToken } from '../entities/access-token'
 
 export const connectMySQL = (): Promise<Connection> => {
@@ -38,7 +37,7 @@ export const connectMySQL = (): Promise<Connection> => {
         // It's bad but we don't have here any strong data in db
         synchronize: true,
         logging: true,
-        entities: [AccessToken, Infractions, MessageRoles, MessageRolesActions],
+        entities: [AccessToken, Infractions],
         charset: 'utf8mb4_unicode_ci',
         cache: {
           type: 'ioredis',
