+++ packages/server/src/lib/stores/index.ts
@@ -3,6 +3,4 @@
 // found in the LICENSE file.
 
 export * from './access-token'
-export * from './help-chan'
 export * from './infractions'
-export * from './message-roles'
